Many overhauls of API so that read/getline make sense
It turns out that I can't read individual characters from a
tty unless I disable line buffering, and that introduces the
problem of needing to process backspace, and combined with
the half-baked concept I had for nonblocking flag, it was all
a bigger mess than it was worth.  I reduced the getline args
to just the file handle, and if the user wants something more
elaborate they can change TTY flags how they want and use the
read method.

Have fairly comprehensive tests now for both read and getline.