Giant overhaul of read/write API
I realized that trying to standardize nonblocking IO and looping
reads until the desired count was a giant undertaking and not
really that valuable to the module, when all a user wants is a
sysread, syswrite, read, and readline from the console that behave
mostly like their normal perl counterparts.

The sysread and syswrite are now the same semantics as perl's own
(aside from IO layers) and getline is renamed append_console_line
to reflect its special purpose (and that it doesn't respect the
perl line separator variable, because it can't without going into
raw console mode) and the background thread that writes a pipe
is given a specially named function write_async, which now has a
promise-like return value that can be used to inspect or wait
on the write operation.