FreeBSD and Win32 compat
FreeBSD ought to be able to run the TTY test of t/20-textline, but
neither writing r or n into the PTY seem to deliver the line
to the TTY.  I should come back to this and figure it out... but
meanwhile a manual test of append_console_line works just fine.

The other mystery is that Devel::CheckLib does not seem to actually
be running the .exe it generates, because it gives false positives
for functions that dont exist, only failing if the *header* doesn't
exist, and the functions that false-positive are in string.h on GNU
which does exist on Win32 but without those functions.