Implement correct ->Inline method
The default Inline method created by ExtUtils::Depends has a few
problems:
  * the "MyModule/Install" directory is not included in INC
  * the typemaps are relative names, but need to be absolute
  * the XS library itself is not referenced in LIBS

That third one might be intentional since normally a XS lib doesn't
export anything other than the boot_ method, but I wish to change
this, since it's the only thing holding back XS libraries from being
used by other XS libraries.

The first two are normally fixed up by the ExtUtils::Depends->load
method, but this process doesn't happen for Inline so Inline gets
an incomplete list.

I've designed sort of a new proposal for ExtUtils::Depends that has
a new "add_exported_c_api" method that lets you declare a function
which should be exported by the .so or .dll, by specifying either
the name or full C prototype.  If this is present, the Inline method
generated by ExtUtils::Depends will include the .so (or .lib or .a)
of the XS shared lib and its path in the LIBS variable, and also
exports the list into FUNCLIST of get_makefile_vars.

This is working on a glibc Linux and MinGW Win32 so far.