New feature-test system
I was getting bogus results with Devel::CheckLib because unless
you specify a list of libs, it doesn't actually try to compile
the example program.  Many of my tests are for things found in
the standard library.

After inspecting implementation of CheckLib and CChecker and
discovering that ExtUtils::CBuilder is a core module, I decided
to just roll my own around that.  I borrowed the feature from
CChecker which generates a header with the macros in it, which
also let me define the headers there and not need them mentioned
in the XS file decorated with a mess of HAVE_x conditionals.