Fix a bug where index() couldn't find a match at the final char, on Win32
This was a bug in my memmem() replacement, which only shows up on
systems lacking that, like Win32.  It was failing in a "safe" manner
by scanning too few characters, but in fixing this it also needs to
be more careful about the arguments it accepts.

I enhanced the unit tests to include more special cases.