New ->unmask_to method, unmask_secrets_to export - #7
These provide new simpler means of revealing the secret to
functions.  The advantage is that they avoid relying on overloaded
stringification of an object, which could possibly confuse the
callee if it was using tests for whether an argment was a ref or a
string.  The argument passed to the function by unmask_to is seen
as a (magical) SvPV, not a blessed object.