New feature - character sets derived from parsing a Regexp
This is a building block for a future 'scan' method and the existing
'index' and future 'rindex' functions.  They will be able to take
a Regexp-ref of exactly one character class and then scan the buffer
for the next occurrence of that character.

This implementation attempts parsing the character class to build a
bitmap of bytes 0x00-0xFF, but if it can't parse or encounters
unicode, it falls back to testing each character with Perl's own
regex engine.

The character sets are created on the fly and then bound to the
Regexp-ref as MAGIC.