Move secret() constructor to XS, and support assigning Spans
Calling secret() was returning a blessed object that lacked the
MAGIC needed, so `$span->copy_to($x= secret)` would fail.  Also
the expression `secret($span)` would fail because I hadn't added a
special case for that yet.

I moved the `assign` implementation to the public API, and also
wrote the constructor and secret() function in XS such that it
always initiates the MAGIC, so now both these scenarios work.