Export C_API::(signature) package variables
This is a bit of an overhaul of the recommended way to access the
C API when not directly depending on the .so

There are now package variables with the full signature of the
function, which lets C code detect them in a single call.  I also
normalized all the signatures, and created a new test to ensure
they don't get changed unintentionally.

I also removed the SECRET_BUFFER_(x)_FUNCTION_POINTERS macros, since
linking directly seems to work, and when linking conditionally a user
is more likely to just want one or two of them instead of importing
all of them.