Improved string matching on Span with multibyte encodings
Previously, Span objects with non-byte encodings required the
pattern to be pre-encoded in the same encoding.  That was kind of
awkward, and unworkable for HEX and BASE64 encodings since they
skip over whitespace.  The new implementation iterates codepoints
of the buffer and compares to a list of codepoints from the pattern
which is decoded from perl scalars as normal (UTF-8 or ISO-8859-1).