Fix MULTIPLICITY support
While I was at it, I decided to take the plunge and define
PERL_NO_GET_CONTEXT and play whack-a-mole with all the places that
need a context passed.

In the process, I discovered that my MAGIC destructor for async
write was declared without pTHX, which seems like it ought to have
been crashing.

I also turned on all warnings and dealt with all the unused variables
and pointer type mismatches gcc could find for me.