Fix Win32 wrt. PERL_NO_GET_CONTEXT
CPAN testers shows a Win32 compie fail because malloc and free were
redefined by Perl to call methods on the Perl interpreter instance.
It seems problematic for an unsanctioned thread (the writer) to hold
a ref to the perl interpreter and call its free method unsynchronized
with the thread of that interpreter, so I'm switching to GlobalAlloc
and GlobalFree Win32 API for the async_result object.