New ConsoleState object for better echo-disable logic
There is a race condition in append_console_line in that the user
(or script, like the tests) may see the prompt and start to reply
to it before the echo gets disabled. To fix this, the prompt needs
to be printed *after* the echo is disabled, and to do that and
accommodate all the scenarios I can think of, there really needs
to be a perl object to disable and restore the echo flag.  This is
the implementation of a scope-guard object for that purpose.

Meanwhile, I also found that TCSAFLUSH was blocking until the pty
holder read from the master, which is why *bsd would fail the test.
Changing that to TCSANOW doesn't block, and kind of fixes the
problem, but reveals the race conditions, which the BSD scheduler
is more likely to trigger than on Linux.